Chore/backend cleanup (#605)
* move dtos to dto folder

* split auth controller by functionnality
