+++ confiture-rest-api/src/audits/audits.controller.ts
@@ -30,12 +30,12 @@ import { CriterionResult } from "src/generated/nestjs-dto/criterionResult.entity
 import { MailService } from "../mail/mail.service";
 import { AuditExportService } from "./audit-export.service";
 import { AuditService } from "./audit.service";
-import { CreateAuditDto } from "./create-audit.dto";
-import { DuplicateAuditDto } from "./duplicate-audit.dto";
-import { UpdateAuditDto } from "./update-audit.dto";
-import { PatchAuditDto } from "./patch-audit.dto";
-import { UpdateResultsDto } from "./update-results.dto";
-import { UploadImageDto } from "./upload-image.dto";
+import { CreateAuditDto } from "./dto/create-audit.dto";
+import { DuplicateAuditDto } from "./dto/duplicate-audit.dto";
+import { UpdateAuditDto } from "./dto/update-audit.dto";
+import { PatchAuditDto } from "./dto/patch-audit.dto";
+import { UpdateResultsDto } from "./dto/update-results.dto";
+import { UploadImageDto } from "./dto/upload-image.dto";
 import { AuthRequired } from "src/auth/auth-required.decorator";
 import { User } from "src/auth/user.decorator";
 import { AuthenticationJwtPayload } from "src/auth/jwt-payloads";
