+++ confiture-rest-api/src/audits/dto/upload-image.dto.ts
