+++ confiture-rest-api/src/audits/reports.controller.ts
@@ -12,7 +12,7 @@ import {
   ApiTags
 } from "@nestjs/swagger";
 
-import { AuditReportDto } from "./audit-report.dto";
+import { AuditReportDto } from "./dto/audit-report.dto";
 import { AuditService } from "./audit.service";
 import { AuditExportService } from "./audit-export.service";
