+++ confiture-rest-api/src/auth/auth.controller.ts
@@ -1,19 +1,14 @@
 import {
   BadRequestException,
   Body,
-  ConflictException,
   Controller,
   Delete,
-  Get,
-  HttpCode,
   Post,
   Put,
-  Query,
   UnauthorizedException
 } from "@nestjs/common";
 import {
   ApiBadRequestResponse,
-  ApiConflictResponse,
   ApiCreatedResponse,
   ApiOkResponse,
   ApiTags,
@@ -24,24 +19,13 @@ import { AuditService } from "../audits/audit.service";
 import { FeedbackService } from "../feedback/feedback.service";
 import { MailService } from "../mail/mail.service";
 import { AuthRequired } from "./auth-required.decorator";
-import {
-  AuthService,
-  InvalidVerificationTokenError,
-  SigninError,
-  TokenRegenerationError,
-  UsernameAlreadyExistsError
-} from "./auth.service";
-import { CreateAccountDto } from "./dto/create-account.dto";
+import { AuthService, SigninError } from "./auth.service";
 import { DeleteAccountResponseDto } from "./dto/delete-account-response.dto";
 import { DeleteAccountDto } from "./dto/delete-account.dto";
-import { ResendVerificationEmailDto } from "./dto/resend-verification-email.dto";
 import { SigninDto } from "./dto/signin.dto";
 import { UpdatePasswordDto } from "./dto/update-password.dto";
-import { VerifyAccountDto } from "./dto/verify-account.dto";
 import { AuthenticationJwtPayload } from "./jwt-payloads";
 import { User } from "./user.decorator";
-import { UpdateEmailDto } from "./update-email.dto";
-import { VerifyEmailUpdateDto } from "./verify-email-update.dto";
 
 @Controller("auth")
 @ApiTags("Authentication")
@@ -52,98 +36,6 @@ export class AuthController {
     private readonly feedback: FeedbackService,
     private readonly audit: AuditService
   ) {}
-
-  /**
-   * Create a new user account.
-   *
-   * The account is not useable right away, it first needs to be verified.
-   * To do this, an email containing a verification link is sent to the user
-   * mail adress (their username).
-   *
-   * This link contains a *verification token* that will be sent to the
-   * `/api/auth/verify` endpoint to activate the account.
-   */
-  @Post("signup")
-  @ApiConflictResponse({
-    description: "A verified account with this username already exists."
-  })
-  @ApiCreatedResponse({
-    description: "Account successfully created (pending verification)."
-  })
-  async createAccount(@Body() body: CreateAccountDto) {
-    try {
-      const verificationToken = await this.auth.createUnverifiedUser(
-        body.username,
-        body.password
-      );
-      await this.email.sendAccountVerificationEmail(
-        body.username,
-        verificationToken
-      );
-    } catch (e) {
-      if (e instanceof UsernameAlreadyExistsError) {
-        throw new ConflictException();
-      }
-      throw e;
-    }
-  }
-
-  @Post("resend-verification-email")
-  @HttpCode(200)
-  @ApiOkResponse({ description: "Verification email has been resent." })
-  @ApiBadRequestResponse({
-    description:
-      "Either no such user exist or the account has already been verified."
-  })
-  async resetVarificationEmail(@Body() body: ResendVerificationEmailDto) {
-    try {
-      const verificationToken = await this.auth.regenerateVerificationToken(
-        body.username
-      );
-      console.log("verificationToken:", verificationToken);
-      await this.email.sendAccountVerificationEmail(
-        body.username,
-        verificationToken
-      );
-    } catch (e) {
-      if (e instanceof TokenRegenerationError) {
-        console.log("Token regeneration failed:", e.message);
-        throw new BadRequestException();
-      }
-      throw e;
-    }
-  }
-
-  /**
-   * Verify an account by validating the given token.
-   */
-  @Post("verify")
-  @HttpCode(200)
-  @ApiOkResponse({ description: "Account successfully verified." })
-  @ApiUnauthorizedResponse({ description: "Invalid verification token." })
-  async verifyAccount(@Body() body: VerifyAccountDto) {
-    try {
-      const email = await this.auth.getEmailFromVerificationToken(body.token);
-
-      await this.auth.verifyAccount(body.token);
-      this.email.sendAccountConfirmationEmail(email).catch((err) => {
-        console.error(`Failed to send email for account creation ${email}`);
-        console.error(err);
-      });
-    } catch (e) {
-      if (e instanceof InvalidVerificationTokenError) {
-        console.log("Account verification failed:", e.message);
-        throw new UnauthorizedException("Invalid token");
-      }
-      throw e;
-    }
-  }
-
-  /** Check if account is verified. */
-  @Get("verified")
-  async isAccountVerified(@Query("username") username: string) {
-    return await this.auth.isAccountVerified(username);
-  }
 
   /**
    * Login using *username and password*. An authentication token is returned.
@@ -246,94 +138,5 @@ export class AuthController {
     });
 
     return;
-  }
-  /** Update the user's email adress. The change is not effective immediately.
-   * The new email adress must first be verified using the
-   * `account/verify-email-update` route. */
-  @Put("account/email")
-  @ApiOkResponse({
-    description:
-      "Email update successfully requested. The new email must be confirmed by using the `account/verify-email-update` route."
-  })
-  @AuthRequired()
-  async updateEmail(
-    @Body() body: UpdateEmailDto,
-    @User() user: AuthenticationJwtPayload
-  ) {
-    if (!(await this.auth.checkCredentialsWithUid(user.sub, body.password))) {
-      throw new UnauthorizedException();
-    }
-    try {
-      const verificationToken = await this.auth.addNewEmail(
-        user.sub,
-        body.newEmail
-      );
-
-      await this.email.sendNewEmailVerificationEmail(
-        body.newEmail,
-        verificationToken
-      );
-    } catch (e) {
-      if (e instanceof UsernameAlreadyExistsError) {
-        throw new ConflictException();
-      }
-      throw e;
-    }
-  }
-
-  /** Trigger a new verification email for the email update. */
-  @Post("account/resend-email-update-verification-email")
-  @HttpCode(200)
-  @AuthRequired()
-  async resendNewEmailVerificationEmail(
-    @User() user: AuthenticationJwtPayload
-  ) {
-    try {
-      const { token, email } =
-        await this.auth.regenerateEmailUpdateVerificationToken(user.sub);
-      await this.email.sendNewEmailVerificationEmail(email, token);
-    } catch (e) {
-      if (e instanceof TokenRegenerationError) {
-        console.log("Token regeneration failed:", e.message);
-        throw new BadRequestException();
-      }
-      throw e;
-    }
-  }
-
-  /** Verify an email adress by receiving the verification token sent the wanted email adress. */
-  @Post("account/verify-email-update")
-  async verifyEmailUpdate(@Body() body: VerifyEmailUpdateDto) {
-    try {
-      const email = await this.auth.getEmailFromVerificationToken(body.token);
-
-      await this.auth.verifyEmailUpdate(body.token);
-      this.email.sendEmailUpdateConfirmationEmail(email);
-    } catch (e) {
-      if (e instanceof InvalidVerificationTokenError) {
-        console.log("Email update verification failed:", e.message);
-        throw new UnauthorizedException("Invalid token");
-      }
-      throw e;
-    }
-  }
-
-  @Post("account/cancel-email-update")
-  @AuthRequired()
-  async cancelEmailUpdate(@User() user: AuthenticationJwtPayload) {
-    await this.auth.cancelEmailUpdate(user.email);
-  }
-
-  /** Checks if given email is verified for the authenticated user. */
-  @Get("account/verified-email-update")
-  @ApiOkResponse({
-    type: Boolean
-  })
-  @AuthRequired()
-  async isNewEmailVerified(
-    @User() user: AuthenticationJwtPayload,
-    @Query("email") email: string
-  ) {
-    return await this.auth.userHasEmail(user.sub, email);
   }
 }
